+++ tests/integration/test_ruvsearch.py
@@ -16,7 +16,9 @@ def test_ruvsearch(tmp_path, testdata):
         model = Model.create_model('mox3.mod')
         remove_covariance_step(model)
         model.datainfo = model.datainfo.derive(path=tmp_path / 'moxo_simulated_resmod.csv')
-        res = run_tool('ruvsearch', model, groups=4, p_value=0.05, skip=[])
+        res = run_tool(
+            'ruvsearch', model, results=model.modelfit_results, groups=4, p_value=0.05, skip=[]
+        )
         iteration = [1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3]
         best_model = retrieve_final_model(res)
         assert (res.cwres_models.index.get_level_values('iteration') == iteration).all()