+++ tests/integration/test_covsearch.py
@@ -23,6 +23,7 @@ def test_default(tmp_path, model_count, start_model):
         res = run_tool(
             'covsearch',
             effects,
+            results=start_model.modelfit_results,
             model=start_model,
         )
 
@@ -53,6 +54,7 @@ def test_default_str(tmp_path, model_count, start_model):
             'LET(CONTINUOUS, [AGE, WT]); LET(CATEGORICAL, SEX)\n'
             'COVARIATE([CL, MAT, VC], @CONTINUOUS, exp, *)\n'
             'COVARIATE([CL, MAT, VC], @CATEGORICAL, cat, *)',
+            results=start_model.modelfit_results,
             model=start_model,
         )